Worked on a few things again
Worked on a few things today, such as the create store orders and create emergency orders use cases, and they are partially complete now. Also, made some updates to the entity and accessor class for the txnitems table, adding a few more properties to this entity class such as item name and item description.