Worked on updating order quantity and order receival
Mainly worked on two things today which are: updating the quantity of an item already in an order as well as the order receival that the warehouse manager can do for orders once submitted. Both of these tasks are largely functioning now. Also, fixed 1 or 2 bugs in the add/edit order form and made a minor update to my SQL script once again.