Continued work on order items mgmt and backorders
Continued work on a few things today, mostly on managing order items and backorders. Believe I fixed a few bugs I found, such as for when updating the quantity of item(s) in an order. Also added two date time pickers (one for the date and one for the time) for the warehouse manager to use. Made a few minor changes to my SQL script too.