Worked on store order use cases
Worked on and have largely completed the three store order use cases for sprint 3: pickup, transport, and deliver store order. Coded and designed a few new forms for this, and had to make some updates to a few DAO classes and to my SQL script for this as well. However, a bit more testing of these use cases wouldn't hurt.