Completed Check Delivery Use Case
Completed the check delivery use case in the desktop portion of this project. Was going to originally do this use case for the web portion, but doing it for this portion ended up making more sense, I think, considering the truck driver can pickup and confirm order deliveries in this desktop portion.